+++ src/test/scala/ai/tripl/arc/validate/SQLValidateSuite.scala
@@ -142,7 +142,7 @@ class SQLValidateSuite extends FunSuite with BeforeAndAfter {
         plugin=new validate.SQLValidate,
         name=testName,
         description=None,
-        inputURI=new URI(testURI),
+        inputURI=Option(new URI(testURI)),
         sql="SELECT true, null",
         sqlParams=Map.empty,
         params=Map.empty
@@ -161,7 +161,7 @@ class SQLValidateSuite extends FunSuite with BeforeAndAfter {
         plugin=new validate.SQLValidate,
         name=testName,
         description=None,
-        inputURI=new URI(testURI),
+        inputURI=Option(new URI(testURI)),
         sql="SELECT true, 'message'",
         sqlParams=Map.empty,
         params=Map.empty
@@ -180,7 +180,7 @@ class SQLValidateSuite extends FunSuite with BeforeAndAfter {
         plugin=new validate.SQLValidate,
         name=testName,
         description=None,
-        inputURI=new URI(testURI),
+        inputURI=Option(new URI(testURI)),
         sql="""SELECT true, '{"stringKey": "stringValue", "numKey": 123}'""",
         sqlParams=Map.empty,
         params=Map.empty
@@ -200,7 +200,7 @@ class SQLValidateSuite extends FunSuite with BeforeAndAfter {
           plugin=new validate.SQLValidate,
           name=testName,
           description=None,
-          inputURI=new URI(testURI),
+          inputURI=Option(new URI(testURI)),
           sql="SELECT false, null",
           sqlParams=Map.empty,
           params=Map.empty
@@ -222,7 +222,7 @@ class SQLValidateSuite extends FunSuite with BeforeAndAfter {
           plugin=new validate.SQLValidate,
           name=testName,
           description=None,
-          inputURI=new URI(testURI),
+          inputURI=Option(new URI(testURI)),
           sql="SELECT false, 'this is my message'",
           sqlParams=Map.empty,
           params=Map.empty
@@ -244,7 +244,7 @@ class SQLValidateSuite extends FunSuite with BeforeAndAfter {
           plugin=new validate.SQLValidate,
           name=testName,
           description=None,
-          inputURI=new URI(testURI),
+          inputURI=Option(new URI(testURI)),
           sql="""SELECT false, TO_JSON(NAMED_STRUCT('stringKey', 'stringValue', 'numKey', 123))""",
           sqlParams=Map.empty,
           params=Map.empty
@@ -267,7 +267,7 @@ class SQLValidateSuite extends FunSuite with BeforeAndAfter {
           plugin=new validate.SQLValidate,
           name=testName,
           description=None,
-          inputURI=new URI(testURI),
+          inputURI=Option(new URI(testURI)),
           sql="SELECT 'string', true",
           sqlParams=Map.empty,
           params=Map.empty
@@ -289,7 +289,7 @@ class SQLValidateSuite extends FunSuite with BeforeAndAfter {
           plugin=new validate.SQLValidate,
           name=testName,
           description=None,
-          inputURI=new URI(testURI),
+          inputURI=Option(new URI(testURI)),
           sql="SELECT true, 'message' WHERE false",
           sqlParams=Map.empty,
           params=Map.empty
@@ -304,7 +304,7 @@ class SQLValidateSuite extends FunSuite with BeforeAndAfter {
           plugin=new validate.SQLValidate,
           name=testName,
           description=None,
-          inputURI=new URI(testURI),
+          inputURI=Option(new URI(testURI)),
           sql="SELECT true, 'message' UNION ALL SELECT true, 'message'",
           sqlParams=Map.empty,
           params=Map.empty
@@ -326,7 +326,7 @@ class SQLValidateSuite extends FunSuite with BeforeAndAfter {
           plugin=new validate.SQLValidate,
           name=testName,
           description=None,
-          inputURI=new URI(testURI),
+          inputURI=Option(new URI(testURI)),
           sql="SELECT true",
           sqlParams=Map.empty,
           params=Map.empty
@@ -341,7 +341,7 @@ class SQLValidateSuite extends FunSuite with BeforeAndAfter {
           plugin=new validate.SQLValidate,
           name=testName,
           description=None,
-          inputURI=new URI(testURI),
+          inputURI=Option(new URI(testURI)),
           sql="SELECT true, 'message', true",
           sqlParams=Map.empty,
           params=Map.empty
@@ -362,7 +362,7 @@ class SQLValidateSuite extends FunSuite with BeforeAndAfter {
         plugin=new validate.SQLValidate,
         name=testName,
         description=None,
-        inputURI=new URI(testURI),
+        inputURI=Option(new URI(testURI)),
         sql="""SELECT 0.1 > ${threshold}, 'message'""",
         sqlParams=Map("threshold" -> "0.05"),
         params=Map.empty
@@ -375,7 +375,7 @@ class SQLValidateSuite extends FunSuite with BeforeAndAfter {
           plugin=new validate.SQLValidate,
           name=testName,
           description=None,
-          inputURI=new URI(testURI),
+          inputURI=Option(new URI(testURI)),
           sql="""SELECT 0.01 > ${threshold}, 'message'""",
           sqlParams=Map("threshold" -> "0.05"),
           params=Map.empty
@@ -397,7 +397,7 @@ class SQLValidateSuite extends FunSuite with BeforeAndAfter {
           plugin=new validate.SQLValidate,
           name=testName,
           description=None,
-          inputURI=new URI(testURI),
+          inputURI=Option(new URI(testURI)),
           sql="SELECT CAST(NULL AS BOOLEAN), CAST(NULL AS STRING)",
           sqlParams=Map.empty,
           params=Map.empty