+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - rename `etl.policy.inlinesql` to `etl.policy.inline.sql` and `ETL_POLICY_INLINESQL` to `ETL_POLICY_INLINE_SQL`.
 - remove forced use of `etl.config.fs.gs.project.id`/`ETL_CONF_GOOGLE_CLOUD_PROJECT_ID` and `etl.config.fs.google.cloud.auth.service.account.json.keyfile`/`ETL_CONF_GOOGLE_CLOUD_AUTH_SERVICE_ACCOUNT_JSON_KEYFILE` to access Google Cloud Storage job files.
 - remove previous optimisation when reading a large number of small `json` files in `JSONExtract`. This is to better align with `DataSourceV2`.
+- added `sql` attribute to `MetadataValidate` allowing inline SQL statements.
 - **BREAKING** disable automatic dropping of unsupported types when performing `*Load` stages (e.g. `ParquetLoad` cannot support `NullType`). Old behavior can be enabled by setting `etl.policy.drop.unsupported`/`ETL_POLICY_DROP_UNSUPPORTED` to `true`.
 - **BREAKING** remove deprecated `etl.config.environment.id` and `ETL_CONF_ENV_ID` in favor of `etl.config.tags` or `ETL_CONF_TAGS`.
 
@@ -121,7 +122,7 @@
 
 - **FIX** defect which reported job failure when running in YARN mode.
 - **FIX** defect relating to Amazon S3 protocol to use when running on [Amazon EMR](https://docs.aws.amazon.com/emr/latest/ManagementGuide/emr-plan-file-systems.html)
-- added  `sql` attribute to `SQLTransform` and `SQLValidate` allowing inline SQL statements.
+- added `sql` attribute to `SQLTransform` and `SQLValidate` allowing inline SQL statements.
 - added capability to replace the returned dataset via a `LifecyclePlugin` `after` hook.
 - removed `tutorials` directory from main arc repo as it is available via [arc-starter](https://github.com/tripl-ai/arc-starter).
 - bump to Spark [2.4.5](https://spark.apache.org/releases/spark-release-2-4-5.html).